refactor: move WS server to outer module
